% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-control.R
\name{add_control}
\alias{add_control}
\title{Add a control to the map}
\usage{
add_control(
  map,
  control_name = c("NavigationControl", "ScaleControl", "FullscreenControl"),
  ...,
  pos = NULL
)
}
\arguments{
\item{map}{A \link{maplibregl} object.}

\item{control_name}{The name of the control.}

\item{...}{options}

\item{pos}{The position of the control.}
}
\description{
Add a control to the map
}
