% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-layer.R
\name{add_layer}
\alias{add_layer}
\title{Add a layer to the map}
\usage{
add_layer(map, style)
}
\arguments{
\item{map}{A \link{maplibregl} object.}

\item{style}{The layers's style.}
}
\description{
Add a layer to the map
}
