% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplibregl.R
\name{maplibregl}
\alias{maplibregl}
\title{Create a maplibregl widget}
\usage{

  maplibregl(style = "https://basemaps.cartocdn.com/gl/dark-matter-gl-style/style.json",
  width = NULL, height = NULL, element_id = NULL, ...)
}
\arguments{
\item{style}{The map's style.}

\item{width}{The width of the widget.}

\item{height}{The height of the widget.}

\item{element_id}{The unique ID of the widget.}

\item{...}{The properties of the map.}
}
\description{
Create a maplibregl widget
}
\examples{
map <- maplibregl(
  center = c(-73.9165, 40.7114),
  zoom = 10,
  pitch = 30,
  element_id = "test"
)

if (interactive()) map
}
