% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sources.R
\name{ml_as_source}
\alias{ml_as_source}
\title{Create a data source}
\usage{
ml_as_source(data, lng = NULL, lat = NULL, ...)
}
\arguments{
\item{data}{A data.frame containing geo coords or an sf object.}

\item{lng, lat}{The column names containing the longitudes and latitudes
if \code{data} is of type \code{data.frame}.}

\item{...}{Optional parameters.}
}
\description{
Create a data source
}
