% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-layer.R
\name{add_layer}
\alias{add_layer}
\title{Add a layer to the map}
\usage{
add_layer(map, style, data = NULL, id = NULL, ...)
}
\arguments{
\item{map}{A \link{maplibregl} object.}

\item{style}{The layers's style.}

\item{data}{(data.frame or sf object, optional) Overwrites the style's \code{source} parameter.}

\item{id}{(character, optional) Overwrites the style's \code{id} parameter.}

\item{...}{Named arguments used to overwrite style parameters.}
}
\description{
Add a layer to the map
}
